+++ packages/dev/devServer/devServerMain.ts
@@ -1,6 +1,7 @@
 import path from "path";
 
 import { startJaviServer } from "^javi/util";
+import { makeJaviDeps } from "^javi/makeJaviDeps";
 import { MainProv, mainWrapper } from "^jab-node";
 
 import conf from "../dev.conf";
@@ -8,6 +9,12 @@ import conf from "../dev.conf";
 const { getPackagePath, projectRoot } = require("../project.conf");
 
 const main = (mainProv: MainProv) => {
+  //default things
+
+  const defaultThings = makeJaviDeps({ projectRoot });
+
+  //start
+
   startJaviServer({
     ...conf,
     name: "Dev",
@@ -24,13 +31,13 @@ const main = (mainProv: MainProv) => {
       absTestFolder: getPackagePath("dev/devServer/testsuite"),
       absTestLogFolder: getPackagePath("dev/devServer/testsuite/_testLogs"),
       tecTimeout: 20000,
+      ...defaultThings.fileService,
     },
 
     jagos: {
       //needed, because dev WatchProcess, Jacs, etc. is used in the subprocesses.
       // so js-scripts can't be started without typescript support.
       alwaysTypeScript: true,
-      projectRoot,
       scriptFolders: [getPackagePath("dev/devServer/scripts")],
       scripts: [
         {
@@ -39,6 +46,7 @@ const main = (mainProv: MainProv) => {
           autoRestart: true,
         },
       ],
+      ...defaultThings.fileService,
     },
 
     makeRoutes: [],