+++ packages/javi/makeJaviDeps.ts
@@ -1,0 +1,22 @@
+import { FileService } from "^jabc";
+import { compareFiles, handleOpenFileInVsCode } from "./util";
+
+export type JaviDefaultDeps = {
+  projectRoot: string;
+};
+
+/**
+ *
+ */
+export const makeJaviDeps = (deps: JaviDefaultDeps) => {
+  const fileService: FileService = {
+    handleOpenFileInEditor: handleOpenFileInVsCode,
+    handleOpenRelativeFileInEditor: (location) =>
+      handleOpenFileInVsCode(location, deps.projectRoot),
+    compareFiles,
+  };
+
+  return {
+    fileService,
+  };
+};