+++ packages/util-javi/index.js
@@ -1,0 +1,1 @@
+throw new Error("util-javi is deprecated.");